requirements.txt
Django==3.2.11
djangorestframework==3.13.1
django-rq==2.5.1
importlib-metadata==4.10.1
Markdown==3.3.6
Pillow==9.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
redis==4.1.1
requests==2.27.1
rq==1.10.1
